Add timestamps to logs issue #47
Fixed loggerDateFormat access modifiers

Fixing the line 106

Fixing the line 106 that could cause some exceptions.

Clear logs. Issue #48

Adding the ability to clear logs. Only a button was added to clear the
textArea.

Check if pdflatex exists. Now in linux as well
Verify the path pointed by the user as pdflatex (Tools -> options -> LaTeX) is valid. If the selected path does not contain pdflatex tool, the user is notified and asked to confirm the choice.

Check if pdflatex exists. Now in linux as well
Verify the path pointed by the user as pdflatex (Tools -> options -> LaTeX) is valid. If the selected path does not contain pdflatex tool, the user is notified and asked to confirm the choice.

Implementing issue #39
The code may look large, but I have made some fixies that had not been seen in #39
Before this commit, if the user click GENERATE, and then CLICK on a pdf file, the Chooser would select "file.pdf". After that, pdflatex would generate the output onto "file.pdf.pdf", instead of overwriting the clicked file. So I fixed this. I remove the ".pdf" if the selected file has this name.

Besides, I implemented the #39 issue. If the output file already exists, the program notify the user and asks to overwrite the file.

Implementing issue #39
Replacing "selected file" with the full path of the file to be overwrited.

Update README.md

Removed planned releases section