Solving the issue #46 Verify the path pointed by the user as pdflatex (Tools -> options -> LaTeX) is valid. If the selected path does not contain pdflatex tool, the user is notified and asked to confirm the choice.
There was a bug in line 103

We have to check if the directory is null before to verify the existence of pdflatex.exe.

Fixing the line 106

Fixing the line 106 that could cause some exceptions.

Clear logs. Issue #48

Adding the ability to clear logs. Only a button was added to clear the
textArea.

Check if pdflatex exists. Now in linux as well
Verify the path pointed by the user as pdflatex (Tools -> options -> LaTeX) is valid. If the selected path does not contain pdflatex tool, the user is notified and asked to confirm the choice.

Check if pdflatex exists. Now in linux as well
Verify the path pointed by the user as pdflatex (Tools -> options -> LaTeX) is valid. If the selected path does not contain pdflatex tool, the user is notified and asked to confirm the choice.