Implementing Issue #35
If the current file has been changed, the file name is changed to "file.tex*".
Besides, when the user clicks on "SAVE", now the file is automatically overwritten (as in most editors). The first "SAVE" click will NOT overwrite the "welcome.tex" file, the program will ask the user a new filename (and path), and then, the next clicks will overwrite that file.
Obviously a new "SAVE AS" button has to be created, in order to make possible to the user save the content in other file.